Attendance accuse entity 수정사항 완료

+++ src/main/java/site/woulduduo/entity/Accuse.java
@@ -17,7 +17,7 @@ public class Accuse {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer accuseNo;
+    private long accuseNo;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="user_account")
